+++ match/tests/Piipan.Match.Core.IntegrationTests/DbFixture.cs
@@ -63,9 +63,10 @@ namespace Piipan.Match.Core.IntegrationTests
                 conn.ConnectionString = ConnectionString;
                 conn.Open();
 
+                conn.Execute("DROP INDEX IF EXISTS index_match_id_on_match_res_events");
+                conn.Execute("DROP TABLE IF EXISTS match_res_events");
                 conn.Execute("DROP TABLE IF EXISTS matches");
                 conn.Execute("DROP TYPE IF EXISTS hash_type");
-                conn.Execute("DROP TYPE IF EXISTS status");
 
                 conn.Close();
             }
@@ -81,9 +82,10 @@ namespace Piipan.Match.Core.IntegrationTests
                 conn.ConnectionString = ConnectionString;
                 conn.Open();
 
+                conn.Execute("DROP INDEX IF EXISTS index_match_id_on_match_res_events");
+                conn.Execute("DROP TABLE IF EXISTS match_res_events");
                 conn.Execute("DROP TABLE IF EXISTS matches");
                 conn.Execute("DROP TYPE IF EXISTS hash_type");
-                conn.Execute("DROP TYPE IF EXISTS status");
                 conn.Execute(sqltext);
 
                 conn.Close();
@@ -98,6 +100,19 @@ namespace Piipan.Match.Core.IntegrationTests
                 conn.Open();
 
                 conn.Execute("DELETE FROM matches");
+
+                conn.Close();
+            }
+        }
+
+        public void ClearMatchResEvents()
+        {
+            using (var conn = Factory.CreateConnection())
+            {
+                conn.ConnectionString = ConnectionString;
+                conn.Open();
+
+                conn.Execute("DELETE FROM match_res_events");
 
                 conn.Close();
             }
@@ -119,7 +134,6 @@ namespace Piipan.Match.Core.IntegrationTests
                         match_id,
                         initiator,
                         states,
-                        status,
                         hash,
                         hash_type,
                         input,
@@ -131,11 +145,39 @@ namespace Piipan.Match.Core.IntegrationTests
                         @MatchId,
                         @Initiator,
                         @States,
-                        @Status::status,
                         @Hash,
                         @HashType::hash_type,
                         @Input::jsonb,
                         @Data::jsonb
+                    )", record);
+
+                conn.Close();
+            }
+        }
+
+        public void InsertMatchResEvent(MatchResEventDbo record)
+        {
+            var factory = NpgsqlFactory.Instance;
+
+            using (var conn = factory.CreateConnection())
+            {
+                conn.ConnectionString = ConnectionString;
+                conn.Open();
+
+                conn.Execute(@"
+                    INSERT INTO match_res_events
+                    (
+                        match_id,
+                        actor,
+                        actor_state,
+                        delta
+                    )
+                    VALUES
+                    (
+                        @MatchId,
+                        @Actor,
+                        @ActorState,
+                        @Delta::jsonb
                     )", record);
 
                 conn.Close();
@@ -160,13 +202,41 @@ namespace Piipan.Match.Core.IntegrationTests
                         hash_type::text,
                         states,
                         input,
-                        data,
-                        invalid,
-                        status::text
+                        data
                     FROM matches
                     WHERE match_id=@MatchId", record);
 
                 result = row.Equals(record);
+
+                conn.Close();
+            }
+
+            return result;
+        }
+
+        public bool HasMatchResEvent(MatchResEventDbo record)
+        {
+            var result = false;
+            var factory = NpgsqlFactory.Instance;
+
+            using (var conn = factory.CreateConnection())
+            {
+                conn.ConnectionString = ConnectionString;
+                conn.Open();
+
+                var row = conn.QuerySingle<MatchResEventDbo>(@"
+                    SELECT
+                        id,
+                        inserted_at,
+                        match_id,
+                        actor,
+                        actor_state,
+                        delta
+                    FROM match_res_events
+                    WHERE id=@Id
+                    ", record);
+
+                result = row.Id.Equals(record.Id);
 
                 conn.Close();
             }