+++ match/tests/Piipan.Match.Func.Api.IntegrationTests/ApiIntegrationTests.cs
@@ -13,6 +13,7 @@ using Newtonsoft.Json;
 using Npgsql;
 using Piipan.Match.Api;
 using Piipan.Match.Api.Models;
+using Piipan.Match.Core.Builders;
 using Piipan.Match.Core.DataAccessObjects;
 using Piipan.Match.Core.Extensions;
 using Piipan.Match.Core.Parsers;
@@ -39,7 +40,7 @@ namespace Piipan.Match.Func.Api.IntegrationTests
                 LdsHash = "eaa834c957213fbf958a5965c46fa50939299165803cd8043e7b1b0ec07882dbd5921bce7a5fb45510670b46c1bf8591bf2f3d28d329e9207b7b6d6abaca5458",
                 CaseId = "CaseIdExample",
                 ParticipantId = "ParticipantIdExample",
-                BenefitsEndDate = new DateTime(1970, 1, 31),
+                ParticipantClosingDate = new DateTime(1970, 1, 15),
                 RecentBenefitMonths = new List<DateTime>() {
                   new DateTime(2021, 5, 31),
                   new DateTime(2021, 4, 30),
@@ -96,6 +97,9 @@ namespace Piipan.Match.Func.Api.IntegrationTests
 
             services.AddTransient<IStreamParser<OrchMatchRequest>, OrchMatchRequestParser>();
 
+            services.AddTransient<IMatchResEventDao, MatchResEventDao>();
+            services.AddTransient<IMatchResAggregator, MatchResAggregator>();
+
             services.AddTransient<IDbConnectionFactory<ParticipantsDb>>(s =>
             {
                 return new BasicPgConnectionFactory<ParticipantsDb>(
@@ -149,7 +153,7 @@ namespace Piipan.Match.Func.Api.IntegrationTests
             Assert.Equal(record.CaseId, person.Matches.First().CaseId);
             Assert.Equal(record.ParticipantId, person.Matches.First().ParticipantId);
             Assert.Equal(state[0], person.Matches.First().State);
-            Assert.Equal(record.BenefitsEndDate, person.Matches.First().BenefitsEndDate);
+            Assert.Equal(record.ParticipantClosingDate, person.Matches.First().ParticipantClosingDate);
             Assert.Equal(record.RecentBenefitMonths, person.Matches.First().RecentBenefitMonths);
             Assert.Equal(record.ProtectLocation, person.Matches.First().ProtectLocation);
             // serialization