+++ query-tool/src/Piipan.QueryTool/Pages/Index.cshtml.cs
@@ -52,7 +52,7 @@ namespace Piipan.QueryTool.Pages
                     {
                         Data = new List<RequestPerson>
                         {
-                            new RequestPerson { LdsHash = digest }
+                            new RequestPerson { LdsHash = digest, CaseId = Query.CaseId, ParticipantId = Query.ParticipantId }
                         }
                     };